refatorado para usar a variável do Facotry como namespace
