+++ src/Parser/RakutenLog/Factory.php
@@ -20,28 +20,15 @@
 namespace Rakuten\Connector\Parser\RakutenLog;
 
 use Rakuten\Connector\Parser\ParserFactory;
-use Rakuten\Connector\Exception\RakutenException;
 
 /**
  * Class Factory
- * @package Rakuten\Connector\Parser\RakuteLog
+ * @package Rakuten\Connector\Parser\RakutePay
  */
 abstract class Factory extends ParserFactory
 {
     /**
-     * @param $class
-     * @return mixed
-     * @throws RakutenException
-     * @throws \ReflectionException
+     * @var string
      */
-    public static function create($class)
-    {
-        if (!class_exists($class)) {
-            throw new RakutenException("Class not Exists in TransactionFactory");
-        }
-
-        $class = self::getClass($class);
-
-        return new $class;
-    }
+    protected static $namespace = "Rakuten\Connector\Parser\RakutenLog";
 }
